Remove pre-merged ZCTA entries once MHAs assigned.
The TopoJSON created by topojson-merge seems to also include the source
geographic feature layer (us_zcta500.geo) in addition to the new layer
it creates (us_dod_mha). But we only need/want the new layer.

I think the intended way to address this is with the ndjson-cli toolset
but I have no clue how it tells which of the JSON objects go on a
newline or not so instead I just make a script to strip out the old
layer in the JSON data.

To avoid new Perl deps I do this in Node.js.  Removing the needless data
shaves off about 60% of the data size, getting us from about 10MB to 4MB
before we do further quantization and potentially compression.