Add a national outline to the map.
This needs a different data set from the Census Bureau because trying to
form this data by running topomerge on the existing ZCTA Census data
gives a misshaped national outline because of the places where a ZCTA
isn't defined.

However the fidelity needed is low so we can get away with the 5m
low-res nation shapefile (which still includes a bunch of things we
don't need). The resulting us_dod_mha.topo.json is now a bit north of 3M
again but the presentation on the web page is much improved in my
opinion.