Fix off-by-one error in showing colors on the map.
The 'selected_cur_rate' field expects the rate table to not use the
first column (which is the MHA ID) but when I changed to embed the rate
table I had stripped this column out. This meant the map was for the
paygrade higher than the one selected.

I should have noticed this by looking at E9 or O10 (which would have
caused an error) but I ended up spotting it due to AR012, which has a
large jump in housing allowance going from E5 to E6 which should have
caused it to stand out on the map visually on the default view (E5), but
it only stuck out when I looked at E4 (and wondered why the E5 -> E4
drop was so prominent).