Use a timer to handle map updates when hovering an MHA.
Now that we hover the current MHA background, it can take a fair bit of
CPU time to do. This causes lags when moving the mouse over the map.

Solve this for now by using a debounce technique until the user has
paused over an MHA and then do all the hard work.