Comment Makefile and rename GeoJSON zcta layer to "us_zcta".
The Makefile now follows an approximately linear sequence going from
final output to intermediate products to source data.

As for the GeoJSON, the layer name we'd been using was an artifact of
ogr2ogr defaults from the base file name, but we should choose an
appropriate name, and now we've done so.